+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [UNRELEASED]
+
+- updated to React 19 for possible performance improvements
+- added better undo last commit support even outside of Aider modes
+- added support for optional autoApprove property in custom commands
+
 ## [0.30.0]
 
 - extended context memory mode for Subagents to Full context or Last message